backend/requirements.txt
pymongo
flask
flask-cors
bcrypt
pyjwt
python-dotenv
pyyaml
qrcode
pillow